comp perfil + editar username + act foto perfil

+++ src/redux/UsuarioDucks.js
@@ -1,5 +1,5 @@
 import firebase from "firebase"
-import { auth } from "../firebase"
+import { auth, db, storage } from "../firebase"
 
 // constantes
 
@@ -43,18 +43,49 @@ export const accesoUsuarioAccion = () => async (dispatch, getState) => {
     try {
         const provider = new firebase.auth.GoogleAuthProvider()
         const res = await auth.signInWithPopup(provider)
-        dispatch({
-            type: ACCESO_USUARIO_EXITO,
-            payload: {
-                user: res.user.email,
-                uid: res.user.uid
-            }
-        })
-        localStorage.setItem('usuario', JSON.stringify({
-            user: res.user.email,
-            uid: res.user.uid
-        }))
+
+        const usuarioExistente = await db.collection('usuarios').doc(res.user.email).get()
+
+        console.log(usuarioExistente)
+
+        if (usuarioExistente.exists) {
+            
+            dispatch({
+                type: ACCESO_USUARIO_EXITO,
+                payload: usuarioExistente.data()
+            })
+
+            localStorage.setItem('usuario', JSON.stringify(usuarioExistente.data()))
+
+        } else {
+
+            await db.collection('usuarios').doc(res.user.email).set({
+                email: res.user.email,
+                uid: res.user.uid,
+                username: res.user.displayName,
+                photo: res.user.photoURL
+            })
+
+            dispatch({
+                type: ACCESO_USUARIO_EXITO,
+                payload: {
+                    email: res.user.email,
+                    uid: res.user.uid,
+                    username: res.user.displayName,
+                    photo: res.user.photoURL
+                }
+            })
+            localStorage.setItem('usuario', JSON.stringify({
+                email: res.user.email,
+                uid: res.user.uid,
+                username: res.user.displayName,
+                photo: res.user.photoURL
+            }))
+
+        }
+
         console.log(res)
+
     } catch (error) {
         dispatch({
             type: ACCESO_USUARIO_ERROR
@@ -82,6 +113,74 @@ export const cerrarSesionAccion = () => async (dispatch) => {
             type: CERRAR_SESION
         })
         localStorage.removeItem('usuario')
+    } catch (error) {
+        console.log(error)
+    }
+
+}
+
+export const editarUsernameAccion = (username) => async (dispatch, getState) => {
+
+    dispatch({
+        type: LOADING
+    })
+
+    const usuario = getState().usuarios.user
+
+    try {
+
+        await db.collection('usuarios').doc(usuario.email).update({
+            username: username
+        })
+
+        dispatch({
+            type: ACCESO_USUARIO_EXITO,
+            payload: {
+                ...usuario, username: username
+            }
+        })
+
+        localStorage.setItem('usuario', JSON.stringify({
+            ...usuario, username: username
+        }))
+
+    } catch (error) {
+
+        console.log(error)
+
+    }
+
+}
+
+export const editarFotoPerfilAccion = (imagen) => async (dispatch, getState) => {
+
+    dispatch({
+        type: LOADING
+    })
+
+    const usuario = getState().usuarios.user
+
+    try {
+
+        const imagenRef = storage.ref().child(usuario.email).child('foto_perfil')
+        await imagenRef.put(imagen)
+        const imagenURL = await imagenRef.getDownloadURL()
+
+        await db.collection('usuarios').doc(usuario.email).update({
+            photo: imagenURL
+        })
+
+        dispatch({
+            type: ACCESO_USUARIO_EXITO,
+            payload: {
+                ...usuario, photo: imagenURL
+            }
+        })
+
+        localStorage.setItem('usuario', JSON.stringify({
+            ...usuario, photo: imagenURL
+        }))
+
     } catch (error) {
         console.log(error)
     }
